Main Menu buttons
This is a C# code for the MainMenu. It contains two buttons which are Play and Exit. For the Play button when the Player clicks it, this will load our first level from our build settings.When the player clicks the Exit button a popup QuitMenu will appear and asks if the player really wants to quit. When the player clicks the "Yes" button this will quit our game. When the player clicks "No" we'll disable the quit menu, meaning it won't be visible anymore enable the Play and Exit buttons again so they can be clicked.