Main Menu buttons
It contains two buttons, the "Play" and "Exit" buttons. When the player clicks the "Play" button  this will load our first level from our build settings. When the Player clicks the "Exit" button a popup Quit Menu will appear thus the Player will choose if he wants to quit the game with a "Yes" or "No". If he clicks "Yes" this will quit our game and exit to windows and if he clicks "No" we'll disable the quit menu, meaning it won't be visible anymore then enable the Play and Exit buttons again so they can be clicked.